+++ WebApp/ToTheRescueWebApplication/ToTheRescueWebApplication/Code/Nodes.cs
@@ -9,7 +9,7 @@ namespace ToTheRescueWebApplication.Code
     {
         public int ID { get; set; }//NodeID
         public int MapID { get; set; }
-        public int XCoordinate { get; set; }
-        public int YCoordinate { get; set; }
+        public double XCoordinate { get; set; }
+        public double YCoordinate { get; set; }
     }
 }
